Add support for store-specific templates
Templates can be set in the admin interface, using the template files
as defaults.

The new renderwithstoretemplate shortcut takes a request object and
template name and combines them in a string of the form
'store-name;template.html', which is then passed along to the normal
render shortcut to be recognized and handled by a new custom template
loader.

This is admittedly pretty hacky, but not terrible in that it avoids
reimplementing too much existing Django code.

Closes #13