Limit store templates
Previously the admin interface would allow the creation of
arbitrarily-named templates. With this change, the overridden templates
are limited to those that are defined in the app itself.

A further improvement would be to pre-populate the new template
definition box with the default value when adding a new store template.